logging util class refactor.
